add access points to allow plugins to define restricted views
